Created an example using BarGraph component
