--- src/components/Graphs/BarGraph.config.js
@@ -1,18 +1,0 @@
-export const properties = {
-    margin: {
-        top: 15,
-        bottom: 20,
-        left: 30,
-        right: 20
-    },
-    padding: 0.1,
-    yTickGrid: true,
-    yTickSizeInner: 6,
-    yTickSizeOuter: 0,
-    xTickGrid: false,
-    xTickSizeInner: 6,
-    xTickSizeOuter: 0,
-    orientation: "vertical",
-    dateHistogram: false,
-    interval: "30s"
-}
