+++ src/components/Graphs/BarGraph/default.config.js
@@ -1,0 +1,3 @@
+export const properties = {
+    orientation: "horizontal",
+}
