+++ src/components/Graphs/BarGraph/styles.css
